move order hash schema to a separate file
