+++ src/contract_wrappers/exchange_wrapper.ts
@@ -45,7 +45,8 @@ import {indexFilterValuesSchema} from '../schemas/index_filter_values_schema';
 import {orderFillRequestsSchema} from '../schemas/order_fill_requests_schema';
 import {orderCancellationRequestsSchema} from '../schemas/order_cancel_schema';
 import {orderFillOrKillRequestsSchema} from '../schemas/order_fill_or_kill_requests_schema';
-import {signedOrderSchema, orderSchema, orderHashSchema} from '../schemas/order_schemas';
+import {orderHashSchema} from '../schemas/order_hash_schema';
+import {signedOrderSchema, orderSchema} from '../schemas/order_schemas';
 import {constants} from '../utils/constants';
 import {TokenWrapper} from './token_wrapper';
 import {decorators} from '../utils/decorators';
