+++ src/schemas/order_schemas.ts
@@ -25,12 +25,6 @@ export const orderSchema = {
     type: 'object',
 };
 
-export const orderHashSchema = {
-    id: '/orderHashSchema',
-    type: 'string',
-    pattern: '^0x[0-9a-fA-F]{64}$',
-};
-
 export const signedOrderSchema = {
     id: '/signedOrderSchema',
     allOf: [
